Keycards, Doors, And Ammo Count
Added colored keycards and doors, if you pick up a keycard it updates the UI to show the player, and if you collide with the door with the same color of keycard it will open the door. I also added a ammo count that cant be restored, so the player has to strategize who they take out and conserve ammo.